Adjusting cors policy
